+++ resolwe_bio/docker_images/dnaseq/packages-manual/primerclip.sh
@@ -5,9 +5,9 @@
 download_and_verify \
     genialis \
     primerclip \
-    v180309 \
-    9496caea64395e5159a0f03abb093a3a79f6d3f0e590ed5abf8af2dabb5682f0 \
-    'https://data2.genialis.com/f/5286d798d4d64acfb5d2/?dl=1#.elf'
+    v171018 \
+    ff118b659e9b40f8847645126b87c158fe04bc7c7f6f709c0ce50c1f0bb46f70 \
+    'https://drive.google.com/uc?id=1xlgMb3aaVUOIMVXqSZIS7ulrOJ3p3BiA#.elf'
 
 add_binary_path \
     genialis \